+++ tests/character_menu_navigation.test.js
@@ -103,6 +103,7 @@ function createHarness() {
     "js/objects/Game_Actor.js",
     "js/objects/Game_Party.js",
     "js/windows/Window_ListViewport.js",
+    "js/windows/Window_TextLayout.js",
     "js/windows/Window_ActorNavigator.js",
     "js/windows/Window_EquipSelect.js",
     "js/windows/Window_Equipment.js",